fix: enhance basic setters
