+++ packages/setting-form/src/index.ts
@@ -3,3 +3,4 @@ export * from './form-item';
 export * from './form-object';
 export * from './form-model';
 export * from './context';
+export * from './setters';
