+++ packages/setting-form/src/setters/index.ts
@@ -1,0 +1,6 @@
+export * from './bool-setter';
+export * from './code-setter';
+export * from './id-setter';
+export * from './number-setter';
+export * from './text-setter';
+export * from './register';
