+++ packages/setting-form/src/setters/text-setter.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import { Input } from 'antd';
+import { FormItemComponentProps } from '../form-item';
+
+export function TextSetter({ onChange, ...rest }: FormItemComponentProps<string>) {
+  return <Input onChange={(e) => onChange?.(e.target.value)} placeholder="请输入文本" {...rest} />;
+}
